refactor: extract hud and restart modules
