+++ src/main.js
@@ -12,6 +12,8 @@ import { startWander } from './entities/wanderers.js';
 
 import { flashBorder, flashFill, blinkButton, applyRandomSkew, setDepthFromBottom, createGrayscaleTexture, createGlowTexture, createHpBar } from './ui/helpers.js';
 
+import { animateStatChange, updateCloudStatus, updateMoneyDisplay, countPrice, cleanupFloatingEmojis } from "./hud.js";
+import { restartGame } from "./endings.js";
 import { keys, requiredAssets, preload as preloadAssets, receipt, emojiFor } from './assets.js';
 import { playOpening, showStartScreen, playIntro } from './intro.js';
 import DesaturatePipeline from './desaturatePipeline.js';
@@ -27,21 +29,12 @@ const DART_MAX_SPEED = (560 / 6) * 3;
 // Also determines where the drink lands when tossed to a customer
 // Lowered by 10px so the drink doesn't land on top of their head
 const DRINK_HOLD_OFFSET = { x: 0, y: -10 };
-
-// Cloud display positions
-// When money reaches $200 the dollar cloud sits at the top value.
-// Hearts use a similar scale based on the MAX_L constant.
-const MONEY_TOP_Y = 5;
-const MONEY_BOTTOM_Y = 35;
-const LOVE_TOP_Y = 5;
-const LOVE_BOTTOM_Y = 35;
 const HEART_EMOJIS = {
   [CustomerState.NORMAL]: null,
-  [CustomerState.BROKEN]: '💔',
-  [CustomerState.MENDING]: '❤️‍🩹',
-  [CustomerState.GROWING]: '💗',
-  [CustomerState.SPARKLING]: '💖',
-  [CustomerState.ARROW]: '💘'
+  [CustomerState.BROKEN]: "💔",
+  [CustomerState.MENDING]: "❤️‍🩹",
+  [CustomerState.GROWING]: "💗",
+  [CustomerState.SPARKLING]: "💖",
 };
 
 // Reactions when a customer receives a drink
@@ -83,163 +76,6 @@ export function setupGame(){
 
 
 
-  function animateStatChange(obj, scene, delta, isLove=false){
-    if(delta===0) return;
-    const up = delta>0;
-    const color = up ? '#0f0' : '#f00';
-    const by = up ? -8 : 8;
-    const originalY = obj.y;
-    const moveDur = isLove && !up ? 160 : 120;
-    scene.tweens.add({targets:obj, y:originalY+by, duration:dur(moveDur), yoyo:true});
-
-    const cloud = isLove ? cloudHeart : cloudDollar;
-    if(cloud){
-      const cOrigY = cloud.y;
-      scene.tweens.add({targets:cloud, y:cOrigY+by, duration:dur(moveDur), yoyo:true});
-    }
-    let on=true;
-    const flashes = isLove && !up ? 4 : 2;
-    const flashDelay = isLove && !up ? 120 : 80;
-    scene.time.addEvent({
-      repeat:flashes,
-      delay:dur(flashDelay),
-      callback:()=>{
-        obj.setColor(on?color:'#fff');
-        if(isLove && !up){
-          obj.setText(String(GameState.love));
-        }
-        on=!on;
-      }
-    });
-    if(cloud){
-      const tint = up ? 0x00ff00 : 0xff0000;
-      let cOn=true;
-      scene.time.addEvent({
-        repeat:flashes,
-        delay:dur(flashDelay),
-        callback:()=>{
-          if(cOn) cloud.setTintFill(tint); else cloud.clearTint();
-          cOn=!cOn;
-        }
-      });
-      scene.time.delayedCall(dur(flashDelay)*(flashes+1)+dur(10),()=>{
-        cloud.clearTint();
-      },[],scene);
-    }
-    scene.time.delayedCall(dur(flashDelay)*(flashes+1)+dur(10),()=>{
-      obj.setColor('#fff');
-      if(isLove && !up){
-        obj.setText(String(GameState.love));
-        // removed wobble animation for the love counter
-      }
-    },[],scene);
-    scene.time.delayedCall(dur(moveDur)*2,()=>{
-      updateCloudStatus(scene);
-    },[],scene);
-  }
-
-  function frameForStat(val){
-    if(val<=0) return 4;
-    if(val===1) return 3;
-    if(val===2) return 2;
-    if(val===3) return 1;
-    return 0;
-  }
-
-  function updateCloudStatus(scene){
-    if(!scene) return;
-    if(cloudHeart) cloudHeart.x = cloudHeartBaseX;
-    if(cloudDollar) cloudDollar.x = cloudDollarBaseX;
-    updateCloudPositions();
-    if(cloudHeart && cloudHeart.setFrame){
-      cloudHeart.setFrame(frameForStat(GameState.love));
-    }
-    if(cloudDollar && cloudDollar.setFrame){
-      cloudDollar.setFrame(frameForStat(Math.floor(GameState.money)));
-    }
-    const amps=[1,3,5,7,10];
-    const durs=[6000,5000,4000,3000,2000];
-    const loveIdx=frameForStat(GameState.love);
-    const moneyIdx=frameForStat(Math.floor(GameState.money));
-    const makeTween=(existing,targets,amp,dur)=>{
-      if(existing){ if(existing.remove) existing.remove(); else if(existing.stop) existing.stop(); }
-      if(!scene.tweens) return null;
-      return scene.tweens.add({targets,x:`+=${amp}`,duration:dur,yoyo:true,repeat:-1,ease:'Sine.easeInOut'});
-    };
-    cloudHeartTween=makeTween(cloudHeartTween,[cloudHeart,loveText],amps[loveIdx],durs[loveIdx]);
-    cloudDollarTween=makeTween(cloudDollarTween,[cloudDollar,moneyText,moneyDollar],amps[moneyIdx],durs[moneyIdx]);
-  }
-
-  function updateCloudPositions(){
-    if(cloudDollar){
-      const ratio=Math.min(MAX_M,Math.max(0,GameState.money))/MAX_M;
-      const newY=MONEY_BOTTOM_Y-(MONEY_BOTTOM_Y-MONEY_TOP_Y)*ratio;
-      cloudDollar.y=newY;
-      const centerX = cloudDollar.x + cloudDollar.displayWidth/2;
-      const centerY = newY + cloudDollar.displayHeight/2;
-      moneyText.setPosition(centerX, centerY);
-      if(moneyDollar){
-        moneyDollar.setPosition(
-          centerX - moneyText.displayWidth/2 - moneyDollar.displayWidth/2,
-          centerY
-        );
-      }
-    }
-    if(cloudHeart){
-      const ratio=Math.min(MAX_L,Math.max(0,GameState.love))/MAX_L;
-      const newY=LOVE_BOTTOM_Y-(LOVE_BOTTOM_Y-LOVE_TOP_Y)*ratio;
-      cloudHeart.y=newY;
-      loveText.setPosition(
-        cloudHeart.x - cloudHeart.displayWidth/2,
-        newY + cloudHeart.displayHeight/2
-      );
-    }
-  }
-
-  function updateMoneyDisplay(){
-    if(!moneyText || !moneyDollar) return;
-    const val = receipt(GameState.money);
-    moneyDollar.setText(val.charAt(0));
-    moneyText.setText(val.slice(1));
-    updateCloudPositions();
-  }
-
-  function countPrice(text, scene, from, to, baseLeft, baseY=15){
-    if(!text || !scene) return;
-    const duration = dur(400);
-    if(scene.tweens && scene.tweens.addCounter){
-      scene.tweens.addCounter({
-        from,
-        to,
-        duration,
-        onUpdate:tween=>{
-          text.setText(receipt(tween.getValue()));
-          text.setPosition(baseLeft + text.displayWidth/2, baseY);
-        },
-        onComplete:()=>{
-          text.setText(receipt(to));
-          text.setPosition(baseLeft + text.displayWidth/2, baseY);
-        }
-      });
-    } else {
-      text.setText(receipt(to));
-      text.setPosition(baseLeft + text.displayWidth/2, baseY);
-    }
-  }
-
-  function cleanupFloatingEmojis(){
-    floatingEmojis.slice().forEach(e=>{
-      if(e && e.destroy) e.destroy();
-      if (typeof removeFloatingEmoji === 'function') {
-        removeFloatingEmoji(e);
-      } else if (GameState && typeof GameState.removeFloatingEmoji === 'function') {
-        GameState.removeFloatingEmoji(e);
-      } else {
-        const i = floatingEmojis.indexOf(e);
-        if(i !== -1) floatingEmojis.splice(i,1);
-      }
-    });
-  }
 
   const HEART_EMOJI_LIST = Object.values(HEART_EMOJIS).filter(Boolean);
 
@@ -568,9 +404,8 @@ export function setupGame(){
   let priceValueYOffset = 15;
   let truck, girl;
   let sideCText;
-  let sideCAlpha=0;
+let sideCAlpha=0;
   let sideCFadeTween=null;
-  let cloudHeartTween=null, cloudDollarTween=null;
   let endOverlay=null;
   // hearts or anger symbols currently animating
 
@@ -4568,89 +4403,6 @@ function dogsBarkAtFalcon(){
     GameState.gameOver=true;
   }
 
-
-  function restartGame(overlay){
-    const scene=this;
-    scene.tweens.killAll();
-    scene.time.removeAllEvents();
-    cleanupFloatingEmojis();
-    cleanupHeartEmojis(scene);
-    cleanupBarks();
-    cleanupBursts();
-    cleanupSparkles(scene);
-    cleanupSparrows(scene);
-    if (GameState.spawnTimer) {
-      GameState.spawnTimer.remove(false);
-      GameState.spawnTimer = null;
-    }
-    if (GameState.dogBarkEvent) {
-      GameState.dogBarkEvent.remove(false);
-      GameState.dogBarkEvent = null;
-    }
-    GameState.falconActive = false;
-    clearDialog.call(scene);
-    dialogDrinkEmoji.attachedTo = null;
-    if(endOverlay){ endOverlay.destroy(); endOverlay=null; }
-    if(sideCText){ sideCText.destroy(); sideCText=null; }
-    reportLine1.setVisible(false);
-    reportLine2.setVisible(false);
-    reportLine3.setVisible(false);
-    tipText.setVisible(false);
-    paidStamp.setVisible(false);
-    lossStamp.setVisible(false);
-    // reset truck and girl to their initial off-screen positions
-    if (truck && girl) {
-      const startX = scene.scale.width + 100;
-      truck.setPosition(startX, 245);
-      girl.setPosition(startX, 245)
-        .setVisible(false)
-        .setAlpha(1);
-      if (girl.clearTint) girl.clearTint();
-    }
-    GameState.money=10.00; GameState.love=3;
-    updateMoneyDisplay();
-    loveText.setText(String(GameState.love));
-    moneyText.setColor('#fff');
-    loveText.setColor('#fff');
-    if (cloudDollar && cloudDollar.clearTint) cloudDollar.clearTint();
-    if (cloudHeart && cloudHeart.clearTint) cloudHeart.clearTint();
-    updateLevelDisplay();
-    updateCloudStatus(scene);
-    if(GameState.activeCustomer){
-      if(GameState.activeCustomer.heartEmoji){ GameState.activeCustomer.heartEmoji.destroy(); GameState.activeCustomer.heartEmoji=null; }
-      GameState.activeCustomer.sprite.destroy();
-    }
-    GameState.activeCustomer=null;
-    cleanupDogs(scene);
-    GameState.queue.forEach(c => {
-      if(c.walkTween){ if(c.walkTween.isPlaying && c.walkTween.stop) c.walkTween.stop(); if(c.walkTween.remove) c.walkTween.remove(); c.walkTween=null; }
-      if(c.heartEmoji){ c.heartEmoji.destroy(); c.heartEmoji=null; }
-      if(c.dog){ if(c.dog.followEvent) c.dog.followEvent.remove(false); c.dog.destroy(); c.dog=null; }
-      if(c.sprite) c.sprite.destroy();
-    });
-    GameState.wanderers.forEach(c => {
-      if(c.walkTween){ if(c.walkTween.isPlaying && c.walkTween.stop) c.walkTween.stop(); if(c.walkTween.remove) c.walkTween.remove(); c.walkTween=null; }
-      if(c.heartEmoji){ c.heartEmoji.destroy(); c.heartEmoji=null; }
-      if(c.dog){ if(c.dog.followEvent) c.dog.followEvent.remove(false); c.dog.destroy(); c.dog=null; }
-      if(c.sprite) c.sprite.destroy();
-    });
-    GameState.queue=[];
-    GameState.wanderers=[];
-    Object.keys(GameState.customerMemory).forEach(k=>{ delete GameState.customerMemory[k]; });
-    GameState.heartWin = null;
-    GameState.servedCount=0;
-    GameState.saleInProgress = false;
-    sideCAlpha=0;
-    sideCFadeTween=null;
-    GameState.gameOver=false;
-    showStartScreen.call(this);
-    scheduleSparrowSpawn(this);
-    if(overlay){
-      scene.time.delayedCall(50,()=>{
-        scene.tweens.add({targets:overlay,alpha:0,duration:400,onComplete:()=>overlay.destroy()});
-      },[],scene);
-    }
-  }
 
    Assets = { keys, requiredAssets, preload };
    Scene = { create, showStartScreen, playIntro };
